+++ changelog.md
@@ -10,6 +10,10 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 * Made `MainChainScripts` in the native token pallet optional. If they are not set, the inherent data
 provider will not query the main chain state or produce inherent data at all.
 * ETCM-8366 - native token management pallet can now observe historical transfers when added after the genesis block
+* Updated to polkadot-stable2409 (aka v1.16.0).
+* * Requires some downstream changes in the node code.
+* * See diff of the commit that adds this changelog line for hints.
+* * Specific changes will depend on the node implementation.
 
 ## Removed